refactor: batch of mechanical Sonar cleanups (dead code, redundant casts/types)
Four independent, zero-behavior-change rule violations swept in one pass:
- S1068 unused private fields (ColumnBuilder.dtype x2, PcoAnsEncoder.sizeLog,
  ZipBombSecurityTest.BOMB — the last a stale leftover from before the tests
  switched to small CI-safe row counts).
- S7475 typed unnamed pattern bindings (`var _`/`boolean _` -> bare `_`).
- S1905 redundant numeric casts made no-ops by an already-wide sibling
  operand (binary numeric promotion already widens to long/double there).
- S5838 AssertJ isEqualTo(0) -> isZero() — applied only where the assertion
  target has a concrete numeric static type; reverted on Number-typed
  targets (Compute.filteredSum's return type) where isZero() doesn't exist
  on ObjectAssert<Number> and Sonar's suggestion doesn't actually compile.

+++ cli/src/main/java/io/github/dfa1/vortex/cli/FilterCommand.java
@@ -255,7 +255,7 @@ final class FilterCommand {
         if (v >= 0) {
             return v;
         }
-        return ((double) (v >>> 1)) * 2.0 + (v & 1L);
+        return (v >>> 1) * 2.0 + (v & 1L);
     }
 
     private static int compareDouble(double colVal, Comparable<?> value) {
